cleaned up functions and moved helper functions to seperate file
